refactor: extract teller libraries, consolidate LZ chain management, and add paginated incentive history
Extract shared logic into TellerWithMultiAssetSupportLib, LayerZeroTellerLib,
and PairwiseRateLimiterLib to reduce code duplication across teller variants.

Consolidate LayerZero chain management: replace allowMessagesFromChain,
allowMessagesToChain, stopMessagesFromChain, stopMessagesToChain, and
setChainGasLimit with addChain (full overwrite) and stopMessages (selective
disable).

Add paginated history retrieval to IncentivePool.

Add comprehensive test coverage for LZ teller rate limiting, principal
history tracking, and reward routing (2,600+ lines).

+++ src/helper/ArcticArchitectureLens.sol
@@ -302,7 +302,7 @@ contract ArcticArchitectureLens {
         if (shareAmount == 0) res.noShares = true;
 
         // Check if user's shares are locked
-        (,,, uint256 shareUnlockTime) = teller.beforeTransferData(account);
+        (,,, uint64 shareUnlockTime) = teller.beforeTransferData(account);
         if (shareUnlockTime > block.timestamp) res.sharesLocked = true;
 
         // Calculate expected assets out
